pyproject.toml
[project]
name = "hr-pulse-backend"
version = "0.1.0"
description = "Backend for HR-Pulse project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-textanalytics>=5.3.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.0",
    "pyodbc>=5.0.0",
    "fastapi>=0.131.0",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.41.0",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "argon2-cffi>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "wordcloud>=1.9.6",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "opentelemetry-instrumentation-httpx>=0.60b1",
]
